+++ src/main/java/dansplugins/factionsystem/CommandInterpreter.java
@@ -2,7 +2,7 @@ package dansplugins.factionsystem;
 
 import dansplugins.factionsystem.commands.*;
 import dansplugins.factionsystem.data.PersistentData;
-import dansplugins.factionsystem.domainobjects.Faction;
+import dansplugins.factionsystem.objects.Faction;
 import dansplugins.factionsystem.utils.Utilities;
 import org.bukkit.ChatColor;
 import org.bukkit.command.CommandSender;
