+++ src/main/java/dansplugins/factionsystem/commands/ForceCommand.java
@@ -3,8 +3,8 @@ package dansplugins.factionsystem.commands;
 import dansplugins.factionsystem.MedievalFactions;
 import dansplugins.factionsystem.StorageManager;
 import dansplugins.factionsystem.data.PersistentData;
-import dansplugins.factionsystem.domainobjects.Faction;
-import dansplugins.factionsystem.domainobjects.PlayerPowerRecord;
+import dansplugins.factionsystem.objects.Faction;
+import dansplugins.factionsystem.objects.PlayerPowerRecord;
 import dansplugins.factionsystem.utils.Utilities;
 import org.bukkit.Bukkit;
 import org.bukkit.ChatColor;
