+++ src/main/java/dansplugins/factionsystem/commands/GateCommand.java
@@ -2,8 +2,8 @@ package dansplugins.factionsystem.commands;
 
 import dansplugins.factionsystem.data.EphemeralData;
 import dansplugins.factionsystem.data.PersistentData;
-import dansplugins.factionsystem.domainobjects.Faction;
-import dansplugins.factionsystem.domainobjects.Gate;
+import dansplugins.factionsystem.objects.Faction;
+import dansplugins.factionsystem.objects.Gate;
 import dansplugins.factionsystem.utils.Utilities;
 import org.bukkit.ChatColor;
 import org.bukkit.command.CommandSender;
