+++ src/main/java/dansplugins/factionsystem/commands/HomeCommand.java
@@ -2,8 +2,8 @@ package dansplugins.factionsystem.commands;
 
 import dansplugins.factionsystem.MedievalFactions;
 import dansplugins.factionsystem.data.PersistentData;
-import dansplugins.factionsystem.domainobjects.ClaimedChunk;
-import dansplugins.factionsystem.domainobjects.Faction;
+import dansplugins.factionsystem.objects.ClaimedChunk;
+import dansplugins.factionsystem.objects.Faction;
 import dansplugins.factionsystem.utils.Utilities;
 import org.bukkit.ChatColor;
 import org.bukkit.Chunk;
