+++ src/main/java/dansplugins/factionsystem/commands/RenameCommand.java
@@ -2,9 +2,9 @@ package dansplugins.factionsystem.commands;
 
 import dansplugins.factionsystem.StorageManager;
 import dansplugins.factionsystem.data.PersistentData;
-import dansplugins.factionsystem.domainobjects.ClaimedChunk;
-import dansplugins.factionsystem.domainobjects.Faction;
-import dansplugins.factionsystem.domainobjects.LockedBlock;
+import dansplugins.factionsystem.objects.ClaimedChunk;
+import dansplugins.factionsystem.objects.Faction;
+import dansplugins.factionsystem.objects.LockedBlock;
 import dansplugins.factionsystem.utils.Utilities;
 import org.bukkit.ChatColor;
 import org.bukkit.command.CommandSender;
