+++ src/main/java/dansplugins/factionsystem/commands/SetHomeCommand.java
@@ -1,8 +1,8 @@
 package dansplugins.factionsystem.commands;
 
 import dansplugins.factionsystem.data.PersistentData;
-import dansplugins.factionsystem.domainobjects.ClaimedChunk;
-import dansplugins.factionsystem.domainobjects.Faction;
+import dansplugins.factionsystem.objects.ClaimedChunk;
+import dansplugins.factionsystem.objects.Faction;
 import dansplugins.factionsystem.utils.Utilities;
 import org.bukkit.ChatColor;
 import org.bukkit.command.CommandSender;
