+++ src/main/java/dansplugins/factionsystem/eventhandlers/BlockInteractionHandler.java
@@ -3,10 +3,10 @@ package dansplugins.factionsystem.eventhandlers;
 import dansplugins.factionsystem.MedievalFactions;
 import dansplugins.factionsystem.data.EphemeralData;
 import dansplugins.factionsystem.data.PersistentData;
-import dansplugins.factionsystem.domainobjects.ClaimedChunk;
-import dansplugins.factionsystem.domainobjects.Faction;
-import dansplugins.factionsystem.domainobjects.Gate;
-import dansplugins.factionsystem.domainobjects.LockedBlock;
+import dansplugins.factionsystem.objects.ClaimedChunk;
+import dansplugins.factionsystem.objects.Faction;
+import dansplugins.factionsystem.objects.Gate;
+import dansplugins.factionsystem.objects.LockedBlock;
 import dansplugins.factionsystem.utils.Utilities;
 import org.bukkit.ChatColor;
 import org.bukkit.Material;
