+++ src/main/java/dansplugins/factionsystem/eventhandlers/DamageEffectsAndDeathHandler.java
@@ -3,9 +3,9 @@ package dansplugins.factionsystem.eventhandlers;
 import dansplugins.factionsystem.MedievalFactions;
 import dansplugins.factionsystem.data.EphemeralData;
 import dansplugins.factionsystem.data.PersistentData;
-import dansplugins.factionsystem.domainobjects.ClaimedChunk;
-import dansplugins.factionsystem.domainobjects.Duel;
-import dansplugins.factionsystem.domainobjects.PlayerPowerRecord;
+import dansplugins.factionsystem.objects.ClaimedChunk;
+import dansplugins.factionsystem.objects.Duel;
+import dansplugins.factionsystem.objects.PlayerPowerRecord;
 import dansplugins.factionsystem.utils.Pair;
 import dansplugins.factionsystem.utils.Utilities;
 import org.bukkit.ChatColor;
