+++ src/main/java/dansplugins/factionsystem/eventhandlers/JoiningLeavingAndSpawningHandler.java
@@ -3,8 +3,8 @@ package dansplugins.factionsystem.eventhandlers;
 import dansplugins.factionsystem.MedievalFactions;
 import dansplugins.factionsystem.data.EphemeralData;
 import dansplugins.factionsystem.data.PersistentData;
-import dansplugins.factionsystem.domainobjects.PlayerActivityRecord;
-import dansplugins.factionsystem.domainobjects.PlayerPowerRecord;
+import dansplugins.factionsystem.objects.PlayerActivityRecord;
+import dansplugins.factionsystem.objects.PlayerPowerRecord;
 import dansplugins.factionsystem.utils.Utilities;
 import org.bukkit.ChatColor;
 import org.bukkit.entity.Monster;
