+++ src/main/java/dansplugins/factionsystem/objects/GateCoord.java
@@ -1,4 +1,4 @@
-package dansplugins.factionsystem.domainobjects;
+package dansplugins.factionsystem.objects;
 
 import org.bukkit.block.Block;
